Adding copied version of spree gift card
Originally came from https://github.com/vinsol-spree-contrib/spree_gift_card
This version worked in old Spree 3.7, but was not compatible with new
Spree 4.4

Took old version so we could start from a base that "worked", then made
changes to make it work with Spree 4.4 and its headless architecture.
As such, this is a work in progress and likely to contain a lot of bloat
at this stage.